feat(registry): configure registry in cmd package
BREAKING CHANGE: add registry.NewRegistry responsible for initialising a
registry with default config.

BREAKING CHANGE: reg.setupRoutes made unexported as it is now
automatically called in NewRegistry.

BREAKING CHANGE: the registry is now only concerned with being a HTTP
request router, and not a server. All server and store configuration
is moved out to the cmd package.