feat: add localfs store
Reads a directory hierarchy of git repositories and exposes them to the
registry. Users are expected to synchronise/update these repositories
outside of the registry itself, e.g., with a cronjob.

This is very similar to the GitHub store, just that it reads the local
filesystem instead of the GitHub API.